Set onclicklistener methods to each individual button instead of needing to use switch case.
Added a way of keeping logged in username when switching pages.

Implemented method for logout button (not sure if it is the correct approach to logging out).

Home screen welcome text changed to welcome logged in user.

Removed top action bar (the big bar which showed app name on top). It was messing up the layout of the settings page.